Documentated exported functions, formatted file, and used the Options class instead of an object..
All method that are use-able by the user have been documented. Hovering on a method inside an IDE will now display the usage of the method, its parameters, and return value.

The last three methods are not documentated yet because theBowja knows better what they do.

Exported all methods/classes have been moved to the bottom of the file to improve readability.

The class 'Options' is now exported to the package so the user can use it as a template for the options parameter.