update index
There are a couple achievements that share the same name and cannot be searched by genshin-db searchFolder function.

Examples:
In English there are two "That's a big CrystalFly"
In Thai there are two อาหารซามูไร.
etc.

The only way to get details about them is by searching by the categories they're in (or 'names') and setting verboseCategories to true.